refactor: extract publish support crate
