+++ Cargo.toml
@@ -83,6 +83,7 @@ monochange_lint = { version = "0.4.0", path = "./crates/monochange_lint" }
 monochange_lint_testing = { version = "0.4.0", path = "./crates/monochange_lint_testing" }
 monochange_linting = { version = "0.4.0", path = "./crates/monochange_linting" }
 monochange_npm = { version = "0.4.0", path = "./crates/monochange_npm" }
+monochange_publish = { version = "0.4.0", path = "./crates/monochange_publish" }
 monochange_python = { version = "0.4.0", path = "./crates/monochange_python" }
 monochange_schema = { version = "0.1.0", path = "./crates/monochange_schema" }
 monochange_semver = { version = "0.4.0", path = "./crates/monochange_semver" }
