+++ crates/monochange/src/lib.rs
@@ -331,7 +331,6 @@ mod release_record;
 mod skill;
 mod subagents;
 mod tracing_setup;
-mod trust_capabilities;
 mod versioned_files;
 mod workspace_ops;
