+++ crates/monochange_publish/readme.md
@@ -1,0 +1,3 @@
+# monochange_publish
+
+Publishing support and trusted-publishing capability helpers for monochange.
