+++ crates/monochange_schema/extract-publish-capabilities.md
@@ -1,0 +1,8 @@
+---
+"monochange": patch
+"monochange_publish": patch
+---
+
+# Extract publish support into a dedicated crate
+
+Move publish report models, built-in command builders, and trusted-publishing capability detection into the new `monochange_publish` crate.
